Move optimizing-type to amoro-iceberg module
