+++ amoro-ams/src/main/java/org/apache/amoro/server/optimizing/plan/PartitionEvaluator.java
@@ -18,7 +18,7 @@
 
 package org.apache.amoro.server.optimizing.plan;
 
-import org.apache.amoro.server.optimizing.OptimizingType;
+import org.apache.amoro.optimizing.OptimizingType;
 import org.apache.iceberg.ContentFile;
 import org.apache.iceberg.DataFile;
 import org.apache.iceberg.PartitionSpec;
