+++ amoro-ams/src/test/java/org/apache/amoro/server/optimizing/TestMixedHiveOptimizing.java
@@ -21,6 +21,7 @@ package org.apache.amoro.server.optimizing;
 import org.apache.amoro.hive.table.SupportHive;
 import org.apache.amoro.io.AuthenticatedHadoopFileIO;
 import org.apache.amoro.io.MixedDataTestHelpers;
+import org.apache.amoro.optimizing.OptimizingType;
 import org.apache.amoro.shade.guava32.com.google.common.collect.Lists;
 import org.apache.amoro.shade.guava32.com.google.common.collect.Streams;
 import org.apache.amoro.table.KeyedTable;
