+++ amoro-ams/src/test/java/org/apache/amoro/server/optimizing/flow/checker/FullOptimizingMove2HiveChecker.java
@@ -19,7 +19,7 @@
 package org.apache.amoro.server.optimizing.flow.checker;
 
 import org.apache.amoro.optimizing.OptimizingInputProperties;
-import org.apache.amoro.server.optimizing.OptimizingType;
+import org.apache.amoro.optimizing.OptimizingType;
 import org.apache.amoro.server.optimizing.RewriteStageTask;
 import org.apache.amoro.server.optimizing.UnKeyedTableCommit;
 import org.apache.amoro.server.optimizing.flow.view.TableDataView;
