+++ amoro-format-iceberg/src/main/java/org/apache/amoro/optimizing/OptimizingType.java
@@ -16,20 +16,10 @@
  * limitations under the License.
  */
 
-package org.apache.amoro.server.optimizing;
+package org.apache.amoro.optimizing;
 
 public enum OptimizingType {
-  MINOR(OptimizingStatus.MINOR_OPTIMIZING),
-  MAJOR(OptimizingStatus.MAJOR_OPTIMIZING),
-  FULL(OptimizingStatus.FULL_OPTIMIZING);
-
-  private final OptimizingStatus status;
-
-  OptimizingType(OptimizingStatus status) {
-    this.status = status;
-  }
-
-  public OptimizingStatus getStatus() {
-    return status;
-  }
+  MINOR,
+  MAJOR,
+  FULL
 }
